Modifié: récupération des images dans firebase en utilisant un service
