+++ src/app/services/firebaseService/firebase-service.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { FirebaseService } from './firebase-service.service';
+
+describe('FirebaseServiceService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: FirebaseService = TestBed.get(FirebaseService);
+    expect(service).toBeTruthy();
+  });
+});
