+++ src/app/services/firebaseService/firebase-service.service.ts
@@ -1,0 +1,34 @@
+import {AngularFireDatabase} from '@angular/fire/database';
+import { AngularFireStorage } from '@angular/fire/storage';
+import {Injectable} from "@angular/core";
+
+@Injectable({
+  providedIn: 'root'
+})
+export class FirebaseService {
+
+  constructor(
+      public afDB: AngularFireDatabase,
+      public afSG: AngularFireStorage
+  ) {
+  }
+
+  getImagesDatabase(getImages) {
+    this.afDB.list('Images/').snapshotChanges(['child_added']).subscribe(images => {
+      images.forEach(image => {
+        this.getImagesStorage(image, getImages);
+      });
+    });
+  }
+
+  getImagesStorage(image: any, images: any) {
+    const imgRef = image.payload.exportVal().ref;
+    this.afSG.ref(imgRef).getDownloadURL().subscribe(imgUrl => {
+      images.push({
+        name: image.payload.exportVal().name,
+        url: imgUrl
+      });
+    });
+  }
+
+}
